server/main_server/requirements.txt
numpy
opencv-python
flask
pillow
requests
gunicorn